[Refactor] Relocate files for CQRS pattern implement
