+++ src/main/java/boombimapi/domain/place/official/entity/OfficialPlace.java
@@ -1,4 +1,4 @@
-package boombimapi.domain.place.domain.entity;
+package boombimapi.domain.place.official.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
