+++ src/main/java/boombimapi/domain/place/official/query/dto/request/ViewportRequest.java
@@ -1,0 +1,10 @@
+package boombimapi.domain.place.official.query.dto.request;
+
+import boombimapi.domain.place.official.query.dto.common.Coordinate;
+
+public record ViewportRequest(
+    Coordinate topLeft,
+    Coordinate bottomRight
+) {
+
+}
