--- src/main/java/boombimapi/domain/place/presentation/dto/request/ViewportRequest.java
@@ -1,10 +1,0 @@
-package boombimapi.domain.place.presentation.dto.request;
-
-import boombimapi.domain.place.presentation.dto.common.Coordinate;
-
-public record ViewportRequest(
-    Coordinate topLeft,
-    Coordinate bottomRight
-) {
-
-}
